refactor shared utils
