+++ packages/shared/package.json
@@ -70,6 +70,10 @@
       "require": "./dist/cjs/utils/translation/index.js",
       "import": "./dist/esm/utils/translation/index.js"
     },
+    "./utils/invoice": {
+      "require": "./dist/cjs/utils/invoice/index.js",
+      "import": "./dist/esm/utils/invoice/index.js"
+    },
     "./utils/*": {
       "require": "./dist/cjs/utils/*.js",
       "import": "./dist/esm/utils/*.js"
